backend/requirements.txt
aiohttp
beautifulsoup4
channels
channels-redis
daphne
django==3.2.10
django-debug-toolbar
djangorestframework
django_cleanup
drf_yasg
m3u8
Pillow
requests
lxml
uwsgi